создает модульные файлы
